requirements.txt
# Core dependencies for PubMed API Scraper
biopython>=1.81
pandas>=2.0.0
openpyxl>=3.1.0  # For Excel export

# Flask web application
Flask>=2.3.0
Werkzeug>=2.3.0

# Optional: Web Scraper (NOT RECOMMENDED)
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Optional: Production deployment
# gunicorn>=21.2.0  # Uncomment for production
# flask-caching>=2.0.0  # Uncomment for caching
# flask-limiter>=3.3.0  # Uncomment for rate limiting
pdfplumber>=0.10.0  # PDF text extraction for Excel full-text export
